+++ src/wokwi_client/__init__.py
@@ -15,7 +15,16 @@ from .__version__ import get_version
 from .client import WokwiClient
 from .client_sync import WokwiClientSync
 from .constants import GET_TOKEN_URL
+from .file_ops import FlashSection, IdfFirmwareUploadResult
 from .vcd import VCDData
 
 __version__ = get_version()
-__all__ = ["WokwiClient", "WokwiClientSync", "VCDData", "__version__", "GET_TOKEN_URL"]
+__all__ = [
+    "WokwiClient",
+    "WokwiClientSync",
+    "FlashSection",
+    "IdfFirmwareUploadResult",
+    "VCDData",
+    "__version__",
+    "GET_TOKEN_URL",
+]